requirements.txt
selenium
behave
pycodestyle
allure-behave
allure-python-commons
PyYAML